[Feature #154240261] Change the way user adds meal
Before this commit, the user experience while adding a new meal was
flawed. The user had to first of all add a meal time, post that to the
back-end, then they would be presented with another form for the rest of
the meal details.

This commit enables users to add all the meal details in one form which
should be more convenient.